Started working on game first game idea
The current game idea is that image objects will move around the screen for 60 seconds, when the user taps on the image it will make it fade away and adds a point to the score. I may change this to make them motionless and have no time limit. This is designed as a distraction game and may not need a time limit and can go on for as long as needed